refactor recipe page parts/landscape logic
