--- frontend/components/Domain/Recipe/RecipePage/RecipePageParts/RecipePageTitleContent.vue
@@ -1,161 +1,0 @@
-<template>
-  <div>
-    <template v-if="!isEditMode && landscape">
-      <v-card-title class="px-0 py-2 ma-0 headline">
-        {{ recipe.name }}
-      </v-card-title>
-      <SafeMarkdown :source="recipe.description" />
-      <div v-if="isOwnGroup" class="pb-2 d-flex justify-center flex-wrap">
-        <RecipeLastMade
-          v-model="recipe.lastMade"
-          :recipe="recipe"
-          class="d-flex justify-center flex-wrap"
-          :class="true ? undefined : 'force-bottom'"
-        />
-      </div>
-      <div class="pb-2 d-flex justify-center flex-wrap">
-        <RecipeTimeCard
-          class="d-flex justify-center flex-wrap"
-          :prep-time="recipe.prepTime"
-          :total-time="recipe.totalTime"
-          :perform-time="recipe.performTime"
-        />
-        <RecipeRating
-          v-if="$vuetify.breakpoint.smAndDown"
-          :key="recipe.slug"
-          v-model="recipe.rating"
-          :recipe-id="recipe.id"
-          :slug="recipe.slug"
-        />
-      </div>
-      <v-divider></v-divider>
-    </template>
-    <template v-else-if="isEditMode">
-      <v-text-field
-        v-model="recipe.name"
-        class="my-3"
-        :label="$t('recipe.recipe-name')"
-        :rules="[validators.required]"
-      />
-      <v-container class="ma-0 pa-0">
-        <v-row>
-          <v-col cols="3">
-            <v-text-field
-              v-model="recipeServings"
-              type="number"
-              :min="0"
-              hide-spin-buttons
-              dense
-              :label="$t('recipe.servings')"
-              @input="validateInput($event, 'recipeServings')"
-            />
-          </v-col>
-          <v-col cols="3">
-            <v-text-field
-              v-model="recipeYieldQuantity"
-              type="number"
-              :min="0"
-              hide-spin-buttons
-              dense
-              :label="$t('recipe.yield')"
-              @input="validateInput($event, 'recipeYieldQuantity')"
-            />
-          </v-col>
-          <v-col cols="6">
-            <v-text-field
-            v-model="recipe.recipeYield"
-            dense
-            :label="$t('recipe.yield-text')"
-          />
-          </v-col>
-        </v-row>
-      </v-container>
-
-      <div class="d-flex flex-wrap" style="gap: 1rem">
-        <v-text-field v-model="recipe.totalTime" :label="$t('recipe.total-time')" />
-        <v-text-field v-model="recipe.prepTime" :label="$t('recipe.prep-time')" />
-        <v-text-field v-model="recipe.performTime" :label="$t('recipe.perform-time')" />
-      </div>
-      <v-textarea v-model="recipe.description" auto-grow min-height="100" :label="$t('recipe.description')" />
-    </template>
-  </div>
-</template>
-
-<script lang="ts">
-import { computed, defineComponent } from "@nuxtjs/composition-api";
-import { useLoggedInState } from "~/composables/use-logged-in-state";
-import { usePageState, usePageUser } from "~/composables/recipe-page/shared-state";
-import { validators } from "~/composables/use-validators";
-import { NoUndefinedField } from "~/lib/api/types/non-generated";
-import { Recipe } from "~/lib/api/types/recipe";
-import RecipeRating from "~/components/Domain/Recipe/RecipeRating.vue";
-import RecipeTimeCard from "~/components/Domain/Recipe/RecipeTimeCard.vue";
-import RecipeLastMade from "~/components/Domain/Recipe/RecipeLastMade.vue";
-
-export default defineComponent({
-  components: {
-    RecipeRating,
-    RecipeTimeCard,
-    RecipeLastMade,
-  },
-  props: {
-    recipe: {
-      type: Object as () => NoUndefinedField<Recipe>,
-      required: true,
-    },
-    landscape: {
-      type: Boolean,
-      default: false,
-    },
-  },
-  setup(props) {
-    const { user } = usePageUser();
-    const { imageKey, isEditMode } = usePageState(props.recipe.slug);
-    const { isOwnGroup } = useLoggedInState();
-
-    const recipeServings = computed<number>({
-      get() {
-        return props.recipe.recipeServings;
-      },
-      set(val) {
-        validateInput(val.toString(), "recipeServings");
-      },
-    });
-
-    const recipeYieldQuantity = computed<number>({
-      get() {
-        return props.recipe.recipeYieldQuantity;
-      },
-      set(val) {
-        validateInput(val.toString(), "recipeYieldQuantity");
-      },
-    });
-
-    function validateInput(value: string | null, property: "recipeServings" | "recipeYieldQuantity") {
-    if (!value) {
-      props.recipe[property] = 0;
-      return;
-    }
-
-    const number = parseFloat(value.replace(/[^0-9.]/g, ""));
-    if (isNaN(number) || number <= 0) {
-      props.recipe[property] = 0;
-      return;
-    }
-
-    props.recipe[property] = number;
-  }
-
-    return {
-      user,
-      imageKey,
-      validators,
-      isEditMode,
-      isOwnGroup,
-      recipeServings,
-      recipeYieldQuantity,
-      validateInput,
-    };
-  },
-});
-</script>
